requirements.txt
streamlit>=1.22.0
opencv-python>=4.7.0
numpy>=1.24.3
Pillow>=9.5.0
tensorflow>=2.12.0
requests>=2.28.2
pandas>=1.5.3
ultralytics>=8.0.0